hack for SetInsertItemsRightToLeft() not being instant
SetInsertItemsLeftToRight() and SetSortBagsRightToLeft() don't happen instantly, thus the UI will retain the old value since the refresh happens too clickly for the "Get" counterparts to reflect the new values. This is a hack wherein we set a 1/2 second timer to refresh the GUI to allow time for the setting to go through.